fix: last row of queue not deleting
If a non-last row was removed it would prevent the
last row from ever being removed. This is because
each "Remove row" button stores what row it is in
with a custom property. If a non-last row was
deleted the last row would hold a value that was
outside of the total number of rows remaining, and
thus could never be deleted. Now I just update the
Row property after a row is removed to fix this.