fix: default to h264 encoding
Before the codec for the blurred video was set
by reading the codec of the source video. This was
causing problems across different systems depending
on the codec of the source and the codecs supported
by the host computer. h264 seems to have broad
compatibility, so just default to use that one.