feat: full rework to use the deface CLI tool
Changed the entire functionality of the app to utilize the very good
deface (https://github.com/ORB-HD/deface) tool for anonymization. Now
the application spawns a subprocess for each video in the queue and runs
the deface application on that video. It is no longer easy (feasible?)
to create an executable for this app, so running the main_window.py
file in a properly configured Python environment is imperative.